zephyr: Refactor the ADC iio_device to more flexible io channels
In order to support write-able voltages, rework the existing ADC
iio_device driver to a more flexible io-channels device that allows a
mix of DAC and ADC channels, appropriately reporting the IIO type
depending on the channel type, and supporting writing raw values to the
DAC channels for setting DAC values.
